Cleaned up code and removed unnecessary comments
-Removed todo comments
-Made demo more user-friendly

+++ stats.h
@@ -5,7 +5,7 @@ std::vector<float> deltaRand(cv::VideoCapture cam, int frames);
 
 //we know that k = 100
 //nums -- a forward list of the FREQUENCY with which numbers show up
-float chiSquare(int* freqs, int frames);
+float chiSquare(int* freqs, float frames);
 
 //swaps two pixels on an image
 void cvSwap(cv::Mat& src, int x1, int y1, int x2, int y2);
